split up the helper class into its constitute parts
