+++ src/kOS.Safe/Compilation/KS/LockFunction.cs
@@ -1,0 +1,14 @@
+using System.Collections.Generic;
+
+namespace kOS.Safe.Compilation.KS
+{
+    public class LockFunction
+    {
+        public List<Opcode> Code { get; private set; }
+
+        public LockFunction()
+        {
+            Code = new List<Opcode>();
+        }
+    }
+}
