+++ src/kOS.Safe/kOS.Safe.csproj
@@ -45,11 +45,17 @@
     <Compile Include="Compilation\CompilerOptions.cs" />
     <Compile Include="Compilation\KS\Compiler.cs" />
     <Compile Include="Compilation\KS\Context.cs" />
-    <Compile Include="Compilation\KS\Helpers.cs" />
+    <Compile Include="Compilation\KS\Subprogram.cs" />
     <Compile Include="Compilation\KS\KSScript.cs" />
+    <Compile Include="Compilation\KS\Lock.cs" />
+    <Compile Include="Compilation\KS\LockCollection.cs" />
+    <Compile Include="Compilation\KS\LockFunction.cs" />
     <Compile Include="Compilation\KS\Parser.cs" />
     <Compile Include="Compilation\KS\ParseTree.cs" />
     <Compile Include="Compilation\KS\Scanner.cs" />
+    <Compile Include="Compilation\KS\SubprogramCollection.cs" />
+    <Compile Include="Compilation\KS\Trigger.cs" />
+    <Compile Include="Compilation\KS\TriggerCollection.cs" />
     <Compile Include="Compilation\Opcode.cs" />
     <Compile Include="Compilation\ProgramBuilder.cs" />
     <Compile Include="Compilation\ProgramBuilderInterpreter.cs" />
